+++ src/lang/modifyAst/transforms.spec.ts
@@ -325,6 +325,47 @@ extrude001 = extrude(profile001, length = 1)`
       expect(newCode).toContain(code + '\n' + expectedNewLine)
     })
 
+    it('should add a standalone translate call on helix selection', async () => {
+      const code = `helix001 = helix(
+  axis = Z,
+  radius = 5,
+  length = 10,
+  revolutions = 5,
+  angleStart = 0,
+)`
+      const ast = assertParse(code, instanceInThisFile)
+      const sourceRange: [number, number, number] = [0, code.length, 0]
+      const helix: Artifact = {
+        type: 'helix',
+        id: 'helix-id',
+        axisId: null,
+        codeRef: {
+          range: sourceRange,
+          pathToNode: getNodePathFromSourceRange(ast, sourceRange),
+          nodePath: { steps: [] },
+        },
+        trajectorySweepId: null,
+        consumed: false,
+      }
+      const artifactGraph: ArtifactGraph = new Map([[helix.id, helix]])
+      const result = addTranslate({
+        ast,
+        artifactGraph,
+        objects: createSelectionFromArtifacts([helix], artifactGraph),
+        x: await getKclCommandValue(
+          '20',
+          instanceInThisFile,
+          rustContextInThisFile
+        ),
+        wasmInstance: instanceInThisFile,
+      })
+      if (err(result)) throw result
+
+      expect(recast(result.modifiedAst, instanceInThisFile)).toContain(
+        `${code}\ntranslate(helix001, x = 20)`
+      )
+    })
+
     it('should push a call in pipe if selection was in variable-less pipe', async () => {
       const code = `startSketchOn(XY)
   |> circle(center = [0, 0], radius = 1)
